requirements.txt
openai-whisper
sentence-transformers
pydub 
ffmpeg 
langchain
langchain_community
langchain_core
langchain_openai
faiss-cpu

